AccountSystemLogic/src/test/java/za/ac/nwu/ac/logic/flow/impl/CreateAccountTransactionFlowImplTest.java
package za.ac.nwu.ac.logic.flow.impl;

import org.junit.After;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.MockitoJUnitRunner;
import za.ac.nwu.ac.domain.dto.AccountTransactionDto;
import za.ac.nwu.ac.domain.dto.AccountTypeDto;
import za.ac.nwu.ac.translator.AccountTransactionTranslator;

import java.time.LocalDate;

import static org.junit.Assert.*;
import static org.mockito.AdditionalAnswers.returnsFirstArg;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.*;

@RunWith(MockitoJUnitRunner.class)
public class CreateAccountTransactionFlowImplTest {
    @Mock
    private AccountTransactionTranslator translator;
    @InjectMocks
    private CreateAccountTransactionFlowImpl flow;

    @Before
    public void setUp() throws Exception {
    }

    @After
    public void tearDown() throws Exception {
    }

    @Test
    public void create() {
        /*when(translator.save(any(AccountTransaction.class))).then(returnsFirstArg());
        AccountTransactionDto result = flow.create(new AccountTransactionDto());
        assertNotNull(result);
        verify(translator, times(1)).save(any(AccountTransactionDto.class));*/
    }
}